# perf(autolinking)!: prioritise files with `*Package.java|*Package.kt` in `getPackageClassName`