--- a/packages/cli-platform-android/src/config/__fixtures__/files/React.kt
+++ b/packages/cli-platform-android/src/config/__fixtures__/files/React.kt
@@ -1,0 +1,26 @@
+/**
+ * Copyright (c) Facebook, Inc. and its affiliates.
+ *
+ * This source code is licensed under the MIT license found in the
+ * LICENSE file in the root directory of this source tree.
+ */
+
+package com.some.example;
+
+import android.view.View
+import com.facebook.react.ReactPackage
+import com.facebook.react.bridge.NativeModule
+import com.facebook.react.bridge.ReactApplicationContext
+import com.facebook.react.uimanager.ReactShadowNode
+import com.facebook.react.uimanager.ViewManager
+import java.util.*
+
+class SomeExampleKotlinPackage : ReactPackage {
+
+    override fun createNativeModules(reactContext: ReactApplicationContext): MutableList<NativeModule>
+            = mutableListOf(MaterialPaletteModule(reactContext))
+
+    override fun createViewManagers(reactContext: ReactApplicationContext?):
+            MutableList<ViewManager<View, ReactShadowNode>> = Collections.emptyList()
+
+}