--- a/packages/cli-platform-android/src/config/isProjectUsingKotlin.ts
+++ b/packages/cli-platform-android/src/config/isProjectUsingKotlin.ts
@@ -1,7 +1,7 @@
 import {getMainActivityFiles} from './findPackageClassName';
 
 export default function isProjectUsingKotlin(sourceDir: string): boolean {
-  const mainActivityFiles = getMainActivityFiles(sourceDir);
+  const mainActivityFiles = getMainActivityFiles(sourceDir, false);
 
   return mainActivityFiles.some((file) => file.endsWith('.kt'));
 }